Fix edge case for new apartment sell by dates
Fixes sell by date calculation in cases where apartment is completed
before it is sold for the first time. The sell by date should be
calculated based on the first sale purchase date in these cases.
