+++ backend/hitas/views/apartment_list.py
@@ -2,9 +2,7 @@ from rest_framework import mixins, viewsets
 
 from hitas.models import Apartment
 from hitas.views.apartment import ApartmentFilterSet, ApartmentListSerializer, ApartmentViewSet
-from hitas.views.utils import (
-    HitasModelMixin,
-)
+from hitas.views.utils import HitasModelMixin
 
 
 class ApartmentListViewSet(HitasModelMixin, mixins.ListModelMixin, viewsets.GenericViewSet):
