+++ backend/hitas/views/apartment_sale.py
@@ -6,9 +6,9 @@ from rest_framework import serializers
 
 from hitas.exceptions import HitasModelNotFound
 from hitas.models import Apartment, ApartmentSale
-from hitas.models.apartment import prefetch_first_sale
-from hitas.models.condition_of_sale import create_conditions_of_sale
 from hitas.models.ownership import Ownership, OwnershipLike, check_ownership_percentages
+from hitas.services.apartment import prefetch_first_sale
+from hitas.services.condition_of_sale import create_conditions_of_sale
 from hitas.utils import lookup_id_to_uuid
 from hitas.views.ownership import OwnershipSerializer
 from hitas.views.utils import HitasModelSerializer, HitasModelViewSet
